refactor(web): decompose 5 large pages to under 400 lines (ADR-006)
Extract inline components from the 5 largest page files:

- channels/page.tsx: 802 → 241 lines
  → ChannelDialog, RuleDialog, ChannelHealthBadge

- chat/page.tsx: 729 → 214 lines
  → useChatPage hook, ChatHeader

- schedules/page.tsx: 606 → 159 lines
  → ScheduleRow, CreateScheduleDialog, scheduleBadges

- settings/page.tsx: 602 → 239 lines
  → ModelConfigRow, AddDynamicProviderForm

- providers/page.tsx: 553 → 302 lines
  → ProviderActivateDialog, TestConnectionButton, DiscoverModelsButton

Total reduction: 3292 → 1155 lines (65%).
Closes #491.
