+++ web/src/app/providers/page.tsx
@@ -1,31 +1,14 @@
-import { useState, useId } from 'react';
-import {
-  Server,
-  Trash2,
-  RefreshCw,
-  Wifi,
-  WifiOff,
-  Globe,
-  HardDrive,
-  Plus,
-  Zap,
-  Search,
-  CheckCircle2,
-  XCircle,
-  Loader2,
-} from 'lucide-react';
+import { useState } from 'react';
+import { Server, Trash2, RefreshCw, Wifi, WifiOff, Globe, HardDrive } from 'lucide-react';
 import { toast } from 'sonner';
 import {
   useProviders,
   useDeleteProvider,
   useDynamicProviderStatuses,
   useProviderTemplates,
-  useAddProvider,
-  useDiscoverModels,
 } from '@/hooks/useProviders';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
 import { Skeleton } from '@/components/ui/skeleton';
 import { EmptyState } from '@/components/EmptyState';
 import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
@@ -38,9 +21,11 @@ import {
   DialogDescription,
   DialogClose,
 } from '@/components/ui/dialog';
+import { ProviderActivateDialog } from '@/components/providers/ProviderActivateDialog';
+import { TestConnectionButton } from '@/components/providers/TestConnectionButton';
+import { DiscoverModelsButton } from '@/components/providers/DiscoverModelsButton';
 import type { ProviderConfig } from '@/lib/api/types';
 import type { ProviderTemplate } from '@/lib/api/providers';
-import { request } from '@/lib/api/client';
 
 function providerIcon(provider?: string) {
   if (!provider) return <Server size={14} className="text-sera-text-muted" />;
@@ -57,251 +42,6 @@ function groupByProvider(providers: ProviderConfig[]) {
     groups[key]!.push(p);
   }
   return groups;
-}
-
-// ── Template Activation Dialog ──────────────────────────────────────────────
-
-function ActivateDialog({
-  template,
-  onClose,
-}: {
-  template: ProviderTemplate | null;
-  onClose: () => void;
-}) {
-  const [apiKey, setApiKey] = useState('');
-  const [baseUrl, setBaseUrl] = useState('');
-  const [selectedModels, setSelectedModels] = useState<Set<string>>(new Set());
-  const addProvider = useAddProvider();
-  const apiKeyId = useId();
-  const baseUrlId = useId();
-
-  if (!template) return null;
-
-  const allSelected = selectedModels.size === template.models.length;
-
-  const handleActivate = async () => {
-    if (!apiKey && !template.baseUrl) {
-      toast.error('API key is required');
-      return;
-    }
-    const models = selectedModels.size > 0 ? Array.from(selectedModels) : template.models;
-    let successCount = 0;
-    for (const modelName of models) {
-      try {
-        await addProvider.mutateAsync({
-          modelName,
-          api: template.api,
-          provider: template.provider,
-          baseUrl: baseUrl || template.baseUrl,
-          apiKey: apiKey || undefined,
-          apiKeyEnvVar: template.apiKeyEnvVar,
-        });
-        successCount++;
-      } catch {
-        toast.error(`Failed to add ${modelName}`);
-      }
-    }
-    if (successCount > 0) {
-      toast.success(
-        `Added ${successCount} model${successCount > 1 ? 's' : ''} from ${template.displayName}`
-      );
-    }
-    onClose();
-  };
-
-  return (
-    <Dialog open={!!template} onOpenChange={(o: boolean) => !o && onClose()}>
-      <DialogContent className="max-w-md">
-        <DialogHeader>
-          <DialogTitle>Activate {template.displayName}</DialogTitle>
-          <DialogDescription>{template.description}</DialogDescription>
-        </DialogHeader>
-        <div className="space-y-4 mt-2">
-          <div>
-            <label htmlFor={apiKeyId} className="text-xs text-sera-text-muted block mb-1">
-              API Key ({template.apiKeyEnvVar})
-            </label>
-            <Input
-              id={apiKeyId}
-              type="password"
-              value={apiKey}
-              onChange={(e) => setApiKey(e.target.value)}
-              placeholder={`Paste your ${template.provider} API key…`}
-            />
-          </div>
-          {!template.baseUrl && (
-            <div>
-              <label htmlFor={baseUrlId} className="text-xs text-sera-text-muted block mb-1">
-                Base URL (optional)
-              </label>
-              <Input
-                id={baseUrlId}
-                value={baseUrl}
-                onChange={(e) => setBaseUrl(e.target.value)}
-                placeholder="https://api.example.com/v1"
-              />
-            </div>
-          )}
-          <div>
-            <div className="flex items-center justify-between mb-2">
-              <label className="text-xs text-sera-text-muted">
-                Models ({selectedModels.size || template.models.length} selected)
-              </label>
-              <button
-                type="button"
-                className="text-[10px] text-sera-accent hover:underline"
-                onClick={() =>
-                  setSelectedModels(allSelected ? new Set() : new Set(template.models))
-                }
-              >
-                {allSelected ? 'Deselect all' : 'Select all'}
-              </button>
-            </div>
-            <div className="max-h-40 overflow-y-auto space-y-1 border border-sera-border rounded-lg p-2">
-              {template.models.map((m) => (
-                <label
-                  key={m}
-                  className="flex items-center gap-2 text-xs text-sera-text cursor-pointer"
-                >
-                  <input
-                    type="checkbox"
-                    checked={selectedModels.has(m) || selectedModels.size === 0}
-                    onChange={(e) => {
-                      const next = new Set(
-                        selectedModels.size === 0 ? template.models : selectedModels
-                      );
-                      if (e.target.checked) next.add(m);
-                      else next.delete(m);
-                      setSelectedModels(next);
-                    }}
-                    className="rounded"
-                  />
-                  <span className="font-mono">{m}</span>
-                </label>
-              ))}
-            </div>
-          </div>
-          <div className="flex gap-3 justify-end">
-            <DialogClose asChild>
-              <Button variant="ghost" size="sm">
-                Cancel
-              </Button>
-            </DialogClose>
-            <Button
-              size="sm"
-              onClick={() => void handleActivate()}
-              disabled={addProvider.isPending}
-            >
-              {addProvider.isPending ? (
-                <Loader2 size={14} className="animate-spin" />
-              ) : (
-                <Plus size={14} />
-              )}
-              Activate
-            </Button>
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-// ── Test Connection Button ──────────────────────────────────────────────────
-
-function TestConnectionButton({ modelName }: { modelName: string }) {
-  const [status, setStatus] = useState<'idle' | 'testing' | 'ok' | 'fail'>('idle');
-
-  const handleTest = async () => {
-    setStatus('testing');
-    try {
-      await request<{ ok: boolean }>(`/providers/${encodeURIComponent(modelName)}/test`, {
-        method: 'POST',
-      });
-      setStatus('ok');
-    } catch {
-      setStatus('fail');
-    }
-    setTimeout(() => setStatus('idle'), 4000);
-  };
-
-  if (status === 'testing')
-    return <Loader2 size={12} className="animate-spin text-sera-text-muted" />;
-  if (status === 'ok') return <CheckCircle2 size={12} className="text-sera-success" />;
-  if (status === 'fail') return <XCircle size={12} className="text-sera-error" />;
-  return (
-    <Tooltip content="Test connection">
-      <button
-        type="button"
-        onClick={() => void handleTest()}
-        className="p-1 text-sera-text-dim hover:text-sera-accent transition-colors"
-        aria-label="Test connection"
-      >
-        <Zap size={12} />
-      </button>
-    </Tooltip>
-  );
-}
-
-// ── Discover Models Button ──────────────────────────────────────────────────
-
-function DiscoverButton({ providerName }: { providerName: string }) {
-  const discover = useDiscoverModels();
-  const addProvider = useAddProvider();
-  const [models, setModels] = useState<string[] | null>(null);
-
-  const handleDiscover = async () => {
-    try {
-      const result = await discover.mutateAsync(providerName);
-      setModels(result.models);
-      toast.success(`Found ${result.models.length} models`);
-    } catch {
-      toast.error('Discovery failed');
-    }
-  };
-
-  const handleAddModel = async (modelName: string) => {
-    try {
-      await addProvider.mutateAsync({
-        modelName,
-        api: 'openai',
-        provider: providerName,
-      });
-      toast.success(`Added ${modelName}`);
-      setModels((prev) => prev?.filter((m) => m !== modelName) ?? null);
-    } catch {
-      toast.error(`Failed to add ${modelName}`);
-    }
-  };
-
-  return (
-    <div>
-      <Button
-        size="sm"
-        variant="ghost"
-        onClick={() => void handleDiscover()}
-        disabled={discover.isPending}
-      >
-        {discover.isPending ? <Loader2 size={12} className="animate-spin" /> : <Search size={12} />}
-        Discover
-      </Button>
-      {models && models.length > 0 && (
-        <div className="bg-sera-surface-active/40 rounded-lg p-2.5 mt-3 ml-4 border border-sera-border/50 space-y-1">
-          {models.map((m) => (
-            <div key={m} className="flex items-center gap-2 text-xs">
-              <span className="font-mono text-sera-text-muted">{m}</span>
-              <button
-                type="button"
-                onClick={() => void handleAddModel(m)}
-                className="text-sera-accent hover:underline text-[10px]"
-              >
-                + Add
-              </button>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
 }
 
 // ── Main Page ───────────────────────────────────────────────────────────────
@@ -410,7 +150,7 @@ export default function ProvidersPage() {
                 <Badge variant="default" className="ml-auto">
                   {models.length} model{models.length !== 1 ? 's' : ''}
                 </Badge>
-                <DiscoverButton providerName={providerName} />
+                <DiscoverModelsButton providerName={providerName} />
               </CardHeader>
               <CardContent className="divide-y divide-sera-border/50">
                 {models.map((m) => {
@@ -519,7 +259,10 @@ export default function ProvidersPage() {
       )}
 
       {/* ── Activate Dialog ──────────────────────────────────────────────── */}
-      <ActivateDialog template={activateTemplate} onClose={() => setActivateTemplate(null)} />
+      <ProviderActivateDialog
+        template={activateTemplate}
+        onClose={() => setActivateTemplate(null)}
+      />
 
       {/* ── Delete Confirmation ──────────────────────────────────────────── */}
       <Dialog open={!!confirmDelete} onOpenChange={(o: boolean) => !o && setConfirmDelete(null)}>
