+++ web/src/app/settings/page.tsx
@@ -1,22 +1,5 @@
 import { useState } from 'react';
-import {
-  Zap,
-  CheckCircle,
-  XCircle,
-  RefreshCw,
-  Radio,
-  Layers,
-  Settings2,
-  Sliders,
-  Activity,
-  Plus,
-  Save,
-  Search,
-  ChevronDown,
-  Brain,
-} from 'lucide-react';
-import { toast } from 'sonner';
-import { updateProviderConfig } from '@/lib/api/providers';
+import { Radio, Layers, Settings2, Sliders, Activity, Plus, Search } from 'lucide-react';
 import {
   useProviders,
   useDynamicProviders,
@@ -25,215 +8,22 @@ import {
   useRemoveDynamicProvider,
 } from '@/hooks/useProviders';
 import { useCircuitBreakers, useResetCircuitBreaker } from '@/hooks/useHealth';
-import * as providersApi from '@/lib/api/providers';
 import { Spinner } from '@/components/ui/spinner';
-import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import { CopyButton } from '@/components/CopyButton';
 import { CloudProviderSection } from '@/components/CloudProviderSection';
 import { DynamicProviderCard } from '@/components/DynamicProviderCard';
 import { GeneralTab } from '@/components/GeneralTab';
 import { CircuitBreakersTab } from '@/components/CircuitBreakersTab';
 import { EmbeddingsTab } from '@/components/EmbeddingsTab';
 import { ErrorBoundary } from '@/components/ErrorBoundary';
+import { ModelConfigRow } from '@/components/ModelConfigRow';
+import { AddDynamicProviderForm } from '@/components/AddDynamicProviderForm';
 
 type Tab = 'providers' | 'models' | 'general' | 'circuit-breakers' | 'embeddings';
-
-// ── Model Config Row ──────────────────────────────────────────────────────────
-
-function ModelConfigRow({
-  model,
-  authStatus,
-  contextWindow,
-  maxTokens,
-  contextStrategy,
-  contextHighWaterMark,
-  reasoning,
-  onSaved,
-}: {
-  model: {
-    modelName: string;
-    provider?: string;
-    api: string;
-    baseUrl?: string;
-    description?: string;
-  };
-  authStatus: string;
-  contextWindow?: number;
-  maxTokens?: number;
-  contextStrategy?: string;
-  contextHighWaterMark?: number;
-  reasoning?: boolean;
-  onSaved: () => void;
-}) {
-  const [expanded, setExpanded] = useState(false);
-  const [saving, setSaving] = useState(false);
-  const [localCtxWindow, setLocalCtxWindow] = useState(String(contextWindow ?? ''));
-  const [localMaxTokens, setLocalMaxTokens] = useState(String(maxTokens ?? ''));
-  const [localStrategy, setLocalStrategy] = useState(contextStrategy ?? 'sliding-window');
-  const [localHighWater, setLocalHighWater] = useState(String(contextHighWaterMark ?? '0.8'));
-  const [localReasoning, setLocalReasoning] = useState(reasoning ?? false);
-
-  const handleSave = async () => {
-    setSaving(true);
-    try {
-      const overrides: Record<string, unknown> = {};
-      if (localCtxWindow) overrides.contextWindow = parseInt(localCtxWindow, 10);
-      if (localMaxTokens) overrides.maxTokens = parseInt(localMaxTokens, 10);
-      if (localStrategy) overrides.contextStrategy = localStrategy;
-      if (localHighWater) overrides.contextHighWaterMark = parseFloat(localHighWater);
-      overrides.reasoning = localReasoning;
-
-      await updateProviderConfig(model.modelName, overrides);
-      toast.success(`Config updated for ${model.modelName}`);
-      onSaved();
-    } catch (err) {
-      toast.error(`Failed: ${err instanceof Error ? err.message : String(err)}`);
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  return (
-    <div className="border border-sera-border rounded-lg overflow-hidden">
-      <button
-        type="button"
-        onClick={() => setExpanded(!expanded)}
-        className="w-full flex items-center justify-between p-3 hover:bg-sera-surface/50 transition-colors text-left"
-      >
-        <div className="flex items-center gap-3">
-          <div className="flex items-center gap-1">
-            <span className="text-sera-text font-mono text-xs">{model.modelName}</span>
-            <CopyButton value={model.modelName} />
-          </div>
-          <span className="text-sera-text-dim text-[10px]">{model.provider ?? ''}</span>
-          <Badge
-            variant={
-              authStatus === 'configured'
-                ? 'success'
-                : authStatus === 'missing'
-                  ? 'warning'
-                  : 'default'
-            }
-            className="text-[9px]"
-          >
-            {authStatus === 'configured'
-              ? 'Active'
-              : authStatus === 'missing'
-                ? 'Key missing'
-                : 'Local'}
-          </Badge>
-          {reasoning && (
-            <span className="flex items-center gap-0.5 text-[9px] text-violet-400">
-              <Brain size={10} /> reasoning
-            </span>
-          )}
-        </div>
-        <ChevronDown
-          size={14}
-          className={`text-sera-text-dim transition-transform ${expanded ? 'rotate-180' : ''}`}
-        />
-      </button>
-
-      {expanded && (
-        <div className="border-t border-sera-border p-4 bg-sera-surface/30 space-y-3">
-          <div className="grid grid-cols-2 lg:grid-cols-3 gap-3">
-            <div>
-              <label className="text-[10px] uppercase text-sera-text-dim block mb-1">
-                Context Window
-              </label>
-              <input
-                type="number"
-                value={localCtxWindow}
-                onChange={(e) => setLocalCtxWindow(e.target.value)}
-                placeholder="128000"
-                className="sera-input w-full text-xs"
-              />
-            </div>
-            <div>
-              <label className="text-[10px] uppercase text-sera-text-dim block mb-1">
-                Max Tokens (output)
-              </label>
-              <input
-                type="number"
-                value={localMaxTokens}
-                onChange={(e) => setLocalMaxTokens(e.target.value)}
-                placeholder="4096"
-                className="sera-input w-full text-xs"
-              />
-            </div>
-            <div>
-              <label className="text-[10px] uppercase text-sera-text-dim block mb-1">
-                Context Strategy
-              </label>
-              <select
-                value={localStrategy}
-                onChange={(e) => setLocalStrategy(e.target.value)}
-                className="sera-input w-full text-xs"
-              >
-                <option value="sliding-window">Sliding Window</option>
-                <option value="summarize">Summarize</option>
-                <option value="truncate">Truncate</option>
-              </select>
-            </div>
-            <div>
-              <label className="text-[10px] uppercase text-sera-text-dim block mb-1">
-                High Water Mark
-              </label>
-              <input
-                type="number"
-                step="0.05"
-                min="0.5"
-                max="0.95"
-                value={localHighWater}
-                onChange={(e) => setLocalHighWater(e.target.value)}
-                className="sera-input w-full text-xs"
-              />
-            </div>
-            <div className="flex items-center gap-2 pt-4">
-              <input
-                type="checkbox"
-                checked={localReasoning}
-                onChange={(e) => setLocalReasoning(e.target.checked)}
-                id={`reasoning-${model.modelName}`}
-                className="accent-sera-accent"
-              />
-              <label htmlFor={`reasoning-${model.modelName}`} className="text-xs text-sera-text">
-                Reasoning model
-              </label>
-            </div>
-          </div>
-          <div className="flex justify-end pt-1">
-            <Button size="sm" onClick={handleSave} disabled={saving}>
-              <Save size={12} className="mr-1" />
-              {saving ? 'Saving...' : 'Save Config'}
-            </Button>
-          </div>
-          <p className="text-[10px] text-sera-text-dim">
-            Context window controls SERA's compaction threshold. The actual model context depends on
-            your provider's settings (e.g., LM Studio model configuration).
-          </p>
-        </div>
-      )}
-    </div>
-  );
-}
 
 function SettingsPageContent() {
   const [tab, setTab] = useState<Tab>('providers');
   const [showAddDynamic, setShowAddDynamic] = useState(false);
-  const [newDynamic, setNewDynamic] = useState({
-    id: '',
-    name: '',
-    baseUrl: 'http://host.docker.internal:1234/v1',
-    apiKey: '',
-  });
-  const [testResult, setTestResult] = useState<{
-    success: boolean;
-    models: string[];
-    error?: string;
-  } | null>(null);
-  const [isTesting, setIsTesting] = useState(false);
 
   const {
     data: providersData,
@@ -336,172 +126,11 @@ function SettingsPageContent() {
                 </div>
 
                 {showAddDynamic && (
-                  <div className="sera-card-static p-5 mb-4 border-sera-accent/30 bg-sera-accent/5 animate-in zoom-in-95 duration-200">
-                    <div className="flex justify-between items-start mb-4">
-                      <h3 className="text-sm font-semibold text-sera-text">
-                        Add LM Studio Instance
-                      </h3>
-                      <button
-                        onClick={() => setShowAddDynamic(false)}
-                        className="text-sera-text-dim hover:text-sera-text"
-                      >
-                        <XCircle size={16} />
-                      </button>
-                    </div>
-
-                    <div className="grid grid-cols-2 gap-4 mb-4">
-                      <div className="space-y-1.5">
-                        <label className="text-[11px] font-medium text-sera-text-dim uppercase tracking-wider">
-                          Provider Name
-                        </label>
-                        <input
-                          type="text"
-                          placeholder="e.g. Local LM Studio"
-                          value={newDynamic.name}
-                          onChange={(e) => setNewDynamic({ ...newDynamic, name: e.target.value })}
-                          className="sera-input text-xs"
-                        />
-                      </div>
-                      <div className="space-y-1.5">
-                        <label className="text-[11px] font-medium text-sera-text-dim uppercase tracking-wider">
-                          Unique ID
-                        </label>
-                        <input
-                          type="text"
-                          placeholder="e.g. lmstudio-1"
-                          value={newDynamic.id}
-                          onChange={(e) =>
-                            setNewDynamic({
-                              ...newDynamic,
-                              id: e.target.value.toLowerCase().replace(/\s+/g, '-'),
-                            })
-                          }
-                          className="sera-input text-xs font-mono"
-                        />
-                      </div>
-                      <div className="space-y-1.5 col-span-2">
-                        <label className="text-[11px] font-medium text-sera-text-dim uppercase tracking-wider">
-                          Base URL (with /v1)
-                        </label>
-                        <input
-                          type="text"
-                          value={newDynamic.baseUrl}
-                          onChange={(e) =>
-                            setNewDynamic({ ...newDynamic, baseUrl: e.target.value })
-                          }
-                          className="sera-input text-xs font-mono"
-                        />
-                        <p className="text-[10px] text-sera-text-dim mt-0.5">
-                          Running in Docker? Use{' '}
-                          <code className="font-mono">host.docker.internal</code> instead of{' '}
-                          <code className="font-mono">localhost</code>
-                        </p>
-                      </div>
-                      <div className="space-y-1.5 col-span-2">
-                        <label className="text-[11px] font-medium text-sera-text-dim uppercase tracking-wider">
-                          API Key <span className="text-sera-text-dim/50">(optional)</span>
-                        </label>
-                        <input
-                          type="password"
-                          value={newDynamic.apiKey}
-                          onChange={(e) => setNewDynamic({ ...newDynamic, apiKey: e.target.value })}
-                          className="sera-input text-xs"
-                        />
-                      </div>
-                    </div>
-
-                    {testResult && (
-                      <div
-                        className={`mb-4 overflow-hidden rounded-lg border text-xs ${
-                          testResult.success
-                            ? 'bg-sera-success/10 border-sera-success/20 text-sera-success'
-                            : 'bg-sera-error/10 border-sera-error/20 text-sera-error'
-                        }`}
-                      >
-                        <div className="p-3 flex items-start gap-2">
-                          {testResult.success ? <CheckCircle size={14} /> : <XCircle size={14} />}
-                          <div>
-                            <p className="font-semibold">
-                              {testResult.success ? 'Connection successful' : 'Connection failed'}
-                            </p>
-                            {!testResult.success && (
-                              <p className="mt-0.5 opacity-90">{testResult.error}</p>
-                            )}
-                            {testResult.success && (
-                              <p className="mt-1 opacity-90">
-                                Found {testResult.models.length} model(s):{' '}
-                                {testResult.models.join(', ')}
-                              </p>
-                            )}
-                          </div>
-                        </div>
-                      </div>
-                    )}
-
-                    <div className="flex gap-3">
-                      <Button
-                        variant="outline"
-                        className="flex-1 text-xs h-10"
-                        disabled={isTesting || !newDynamic.baseUrl}
-                        onClick={async () => {
-                          setIsTesting(true);
-                          setTestResult(null);
-                          try {
-                            const res = await providersApi.testDynamicConnection(
-                              newDynamic.baseUrl,
-                              newDynamic.apiKey
-                            );
-                            setTestResult(res);
-                          } catch (err: unknown) {
-                            setTestResult({
-                              success: false,
-                              models: [],
-                              error: err instanceof Error ? err.message : String(err),
-                            });
-                          } finally {
-                            setIsTesting(false);
-                          }
-                        }}
-                      >
-                        {isTesting ? (
-                          <RefreshCw className="animate-spin" size={14} />
-                        ) : (
-                          <Zap size={14} />
-                        )}
-                        Test & Discover
-                      </Button>
-                      <Button
-                        className="flex-1 text-xs bg-sera-accent hover:bg-sera-accent-hover text-sera-bg h-10"
-                        disabled={
-                          !(
-                            newDynamic.name &&
-                            newDynamic.id &&
-                            newDynamic.baseUrl &&
-                            testResult?.success &&
-                            !addDynamic.isPending
-                          )
-                        }
-                        onClick={() => {
-                          addDynamic.mutate({
-                            ...newDynamic,
-                            type: 'lm-studio',
-                            enabled: true,
-                            intervalMs: 60000,
-                          });
-                          setShowAddDynamic(false);
-                          setNewDynamic({
-                            id: '',
-                            name: '',
-                            baseUrl: 'http://host.docker.internal:1234/v1',
-                            apiKey: '',
-                          });
-                          setTestResult(null);
-                        }}
-                      >
-                        <Save size={14} /> Save Provider
-                      </Button>
-                    </div>
-                  </div>
+                  <AddDynamicProviderForm
+                    onClose={() => setShowAddDynamic(false)}
+                    onSuccess={() => void refetchDynamic()}
+                    addDynamic={addDynamic}
+                  />
                 )}
 
                 <div className="grid grid-cols-1 lg:grid-cols-2 gap-4">
